@@ -14,7 +14,7 @@ from pyiron_workflow import as_function_node
 
 @as_function_node("fig")
 def Scatter(
-        x: Optional[list | np.ndarray] = None, y: Optional[list | np.ndarray] = None
+    x: Optional[list | np.ndarray] = None, y: Optional[list | np.ndarray] = None
 ):
     from matplotlib import pyplot as plt
 
@@ -40,13 +40,13 @@ def Histogram(x: Optional[list | np.ndarray], bins: int = 50):
 
 @as_function_node("axis")
 def Plot(
-        x: Optional[list | np.ndarray | pandas.core.series.Series] = None,
-        y: Optional[list | np.ndarray | pandas.core.series.Series] = None,
-        axis: Optional[object] = None,
-        title: Optional[str] = '',
-        color: Optional[str] = 'b',
-        symbol: Optional[str] = 'o',
-        legend_label: Optional[str] = ''
+    x: Optional[list | np.ndarray | pandas.core.series.Series] = None,
+    y: Optional[list | np.ndarray | pandas.core.series.Series] = None,
+    axis: Optional[object] = None,
+    title: Optional[str] = "",
+    color: Optional[str] = "b",
+    symbol: Optional[str] = "o",
+    legend_label: Optional[str] = "",
 ):
     from matplotlib import pyplot as plt
 
@@ -62,10 +62,10 @@ def Plot(
 
 @as_function_node("linspace")
 def Linspace(
-        start: Optional[int | float] = 0.0,
-        stop: Optional[int | float] = 1.0,
-        num: Optional[int] = 50,
-        endpoint: Optional[bool] = True
+    start: Optional[int | float] = 0.0,
+    stop: Optional[int | float] = 1.0,
+    num: Optional[int] = 50,
+    endpoint: Optional[bool] = True,
 ):
     from numpy import linspace
 
@@ -73,18 +73,16 @@ def Linspace(
 
 
 @as_function_node("mean")
-def Mean(
-        numbers: Optional[list | np.ndarray] = None
-):
+def Mean(numbers: Optional[list | np.ndarray] = None):
     return np.mean(numbers)
 
 
 @as_function_node("axes")
 def Subplot(
-        nrows: Optional[int] = 1,
-        ncols: Optional[int] = 1,
-        sharex: Optional[bool] = False,
-        sharey: Optional[bool] = False
+    nrows: Optional[int] = 1,
+    ncols: Optional[int] = 1,
+    sharex: Optional[bool] = False,
+    sharey: Optional[bool] = False,
 ):
     from matplotlib import pyplot as plt
 
@@ -93,11 +91,9 @@ def Subplot(
 
 
 @as_function_node("axis")
-def Title(
-        axis: Optional[object] = None,
-        title: Optional[str] = ''
-):
+def Title(axis: Optional[object] = None, title: Optional[str] = ""):
     from matplotlib import pyplot as plt
+
     if axis is None:
         axis = plt
 
